Updated java classes that use the XYArray

+++ ui/org.omnetpp.common.core/src/org/omnetpp/common/color/ColorFactory.java
@@ -853,7 +853,7 @@ public class ColorFactory {
     /**
      * Returns a "good" light color.
      */
-    public static Color getGoodLightColor(int i) {
+    public static Color getGoodLightColor(long i) {
         return goodLightColors[i % goodLightColors.length];
     }
